🚚(frontend) move SaleTunnel into components
This component will be use in the learner dashboard enrollment listing.
