+++ src/frontend/js/components/PaymentButton/_styles.scss
