+++ src/frontend/js/components/SaleTunnel/components/RegisteredCreditCard/_styles.scss
