+++ src/frontend/js/components/SaleTunnel/components/RegisteredCreditCard/index.spec.tsx
