+++ src/frontend/js/components/SaleTunnel/components/SaleTunnelStepPayment/_styles.scss
