+++ src/frontend/js/components/SaleTunnel/components/SaleTunnelStepResume/_styles.scss
