+++ src/frontend/js/components/SaleTunnel/components/SaleTunnelStepResume/index.spec.tsx
