+++ src/frontend/js/components/SaleTunnel/components/SaleTunnelStepValidation/_styles.scss
