+++ src/frontend/js/components/SaleTunnel/components/SaleTunnelStepValidation/index.spec.tsx
