+++ src/frontend/js/components/SaleTunnel/components/SaleTunnelStepValidation/index.tsx
