+++ src/frontend/js/hooks/useStepManager/index.spec.ts
